+++ __tests__/commands/versions/update.test.ts
@@ -5,7 +5,7 @@ import { describe, beforeAll, afterEach, it, expect } from 'vitest';
 import Command from '../../../src/commands/versions/update.js';
 import { APIv1Error } from '../../../src/lib/apiError.js';
 import { getAPIV1Mock } from '../../helpers/get-api-mock.js';
-import { runCommand } from '../../helpers/setup-oclif-config.js';
+import { runCommand } from '../../helpers/oclif.js';
 
 const key = 'API_KEY';
 const version = '1.0.0';
