+++ __tests__/commands/whoami.test.ts
@@ -3,7 +3,7 @@ import { describe, afterEach, it, expect, beforeAll } from 'vitest';
 import pkg from '../../package.json';
 import Command from '../../src/commands/whoami.js';
 import configStore from '../../src/lib/configstore.js';
-import { runCommand } from '../helpers/setup-oclif-config.js';
+import { runCommand } from '../helpers/oclif.js';
 
 describe('rdme whoami', () => {
   let run: (args?: string[]) => Promise<string>;
