+++ __tests__/lib/createGHA.test.ts
@@ -14,7 +14,7 @@ import { getMajorPkgVersion } from '../../src/lib/getPkgVersion.js';
 import { after, before } from '../helpers/get-gha-setup.js';
 import getGitRemoteMock from '../helpers/get-git-mock.js';
 import ghaWorkflowSchema from '../helpers/github-workflow-schema.json' with { type: 'json' };
-import { setupOclifConfig } from '../helpers/setup-oclif-config.js';
+import { setupOclifConfig } from '../helpers/oclif.js';
 
 const testWorkingDir = process.cwd();
